+++ lib/src/analyzers/lint_analyzer/rules/models/rule_type.dart
@@ -8,4 +8,5 @@ class RuleType {
   static const flutter = RuleType._('flutter');
   static const intl = RuleType._('intl');
   static const angular = RuleType._('angular');
+  static const flame = RuleType._('flame');
 }